verify that req.user.id is not null add ghost icon if ghost_address === true Add italic message in profile to indicate user has to change his address

+++ server/routes/updateAddress.ts
@@ -14,41 +14,52 @@ const updateAddress = async (models: DB, req: Request, res: Response, next: Next
     if (!chain) {
         return next(new Error(Errors.NoChain));
     }
-    const { id: ghostAddressId } = await models.Address.scope('withPrivateData').findOne({
-        where: { chain, ghost_address: true, user_id: req.user.id }
-    }) || {};
-
-    const { id: newAddressId } = await models.Address.scope('withPrivateData').findOne({
-        where: { chain, user_id: req.user.id, address }
-    }) || {};
-
+    const transaction = await sequelize.transaction();
     try {
-        if (ghostAddressId && newAddressId) {
-            const transaction = await sequelize.transaction();
-            // update address in comments
-            await models.OffchainComment.update({ address_id: newAddressId }, { where: { address_id: ghostAddressId }, transaction });
+        if (req.user.id) {
+            const { id: ghostAddressId } = await models.Address.scope('withPrivateData').findOne({
+                where: { chain, ghost_address: true, user_id: req.user.id }
+            }) || {};
 
-            // update address in reactions
-            await models.OffchainReaction.update({ address_id: newAddressId }, { where: { address_id: ghostAddressId }, transaction });
+            const { id: newAddressId } = await models.Address.scope('withPrivateData').findOne({
+                where: { chain, user_id: req.user.id, address }
+            }) || {};
 
-            // update address in threads
-            await models.OffchainThread.update({ address_id: newAddressId }, { where: { address_id: ghostAddressId }, transaction });
+            if (ghostAddressId && newAddressId) {
+                // update address in comments
+                await models.OffchainComment.update({ address_id: newAddressId }, { where: { address_id: ghostAddressId }, transaction });
 
-            // update address in roles
-            await models.Role.update({ address_id: newAddressId }, { where: { address_id: ghostAddressId }, transaction });
+                // update address in reactions
+                await models.OffchainReaction.update({ address_id: newAddressId }, { where: { address_id: ghostAddressId }, transaction });
 
-            // update address in profile
-            await models.OffchainProfile.update({ address_id: newAddressId }, { where: { address_id: ghostAddressId }, transaction });
+                // update address in threads
+                await models.OffchainThread.update({ address_id: newAddressId }, { where: { address_id: ghostAddressId }, transaction });
 
-            // delete ghost address from Address
-            await models.Address.destroy({ where: { id: ghostAddressId }, transaction });
+                // update address in roles
+                await models.Role.update({ address_id: newAddressId }, { where: { address_id: ghostAddressId }, transaction });
 
-            return res.json({ status: 'Success', result: 'Ghost Address has been successfully replaced and deleted in all tables' });
+                // update address in profile
+                await models.OffchainProfile.update({ address_id: newAddressId }, { where: { address_id: ghostAddressId }, transaction });
+
+                // delete role by address
+                await models.Role.destroy({ where: { address_id: ghostAddressId }, transaction });
+
+                // delete ghost address from Address
+                await models.Address.destroy({ where: { id: ghostAddressId }, transaction });
+                await transaction.commit()
+                return res.json({
+                    success: true,
+                    ghostAddressId,
+                    newAddressId,
+                    result: 'Ghost Address has been successfully replaced and deleted in all tables'
+                });
+            }
         }
-        return res.json({}) ;
+        return res.json('user id or ghost address or new address not found') ;
     } catch (e) {
-        return res.json({ status: 'Error', result: e });
+        await transaction.rollback();
         console.log(e);
+        return res.json({ status: 'Error', result: e });
     }
 }
